feat: refactor effective set handling with new EffectiveSetCalculator service
