+++ envgene-inventory-service/src/main/java/org/qubership/colly/db/data/ParamsetContext.java
@@ -1,7 +1,24 @@
 package org.qubership.colly.db.data;
 
 public enum ParamsetContext {
-    DEPLOYMENT,
-    RUNTIME,
-    PIPELINE
+    DEPLOYMENT("deployment"),
+    RUNTIME("runtime"),
+    PIPELINE("pipeline");
+
+    private final String key;
+
+    ParamsetContext(String key) {
+        this.key = key;
+    }
+
+    public String key() {
+        return key;
+    }
+
+    public static ParamsetContext fromKey(String key) {
+        for (ParamsetContext ctx : values()) {
+            if (ctx.key.equals(key)) return ctx;
+        }
+        return null;
+    }
 }
